requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-multipart
google-genai
pydantic
python-dotenv